int -> int64_t in seg_tree::find_last

+++ library/data_structures/seg_tree_uncommon/find_last.hpp
@@ -10,10 +10,10 @@
  * @time O(log(n))
  * @space O(log(n)) for recursion stack; no new nodes are allocated
  */
-int find_last(int le, int ri, function<bool(int, int, int)> f) {
+int find_last(int le, int ri, function<bool(int64_t, int, int)> f) {
 	return find_last(le, ri, f, 0, n, 1);
 }
-int find_last(int le, int ri, const function<bool(int, int, int)>& f, int tl,
+int find_last(int le, int ri, const function<bool(int64_t, int, int)>& f, int tl,
               int tr, int u) {
 	if (ri <= tl || tr <= le) return le - 1;
 	if (tr - tl == 1) {
